server/requirements_clean.txt
# Core FastAPI dependencies
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
python-multipart==0.0.21

# Authentication and database
supabase==2.10.0
email-validator==2.1.1

# OCR and image processing
pytesseract==0.3.13
Pillow==12.0.0

# Optional AI enhancement (can be omitted if not needed)
# openai==1.54.5

# Core Python dependencies (usually included)
pydantic==2.12.5
typing-extensions==4.15.0